Don't attempt to vectorize complex loads.
In principle we could do this, but we need some fixes to
our vector size heuristics first. Right now, we generate
loops of size 4, which is too much (1024 bytes per warp).
But I'm not sure it's actually helpful in any case.

This fixes the `error: 'arith.constant' op requires attribute
'value'` errors.

PiperOrigin-RevId: 641183284