[XLA:FFI] Move stream insertion operator for C API datatypes to global namespace.
In case this operator is called by a user using a macro, it happens in the global namespace. If the stream insertion operator is not found there, ADL never looks in `xla::ffi` namespace (because C API enums are defined in global namespace), so the generic `std` version for int is found instead. It results in cryptic messages like 'Wrong scalar data type: expected 4 but got 5'.

Moving stream insertion operator to global namespace solves this problem.

PiperOrigin-RevId: 642625139